fix: Resolve 422 Unprocessable Entity block during CSV export
- Updated Pydantic typing in schemas.py to explicitly allow fractional floats for radius_m and azimuth, preventing validation failures when importing/exporting High Gain (*1.2) sectors.

+++ temp_test_zip/schemas.py
@@ -6,8 +6,8 @@ class SiteEditRequest(BaseModel):
     type: str
     lat: float
     lon: float
-    azimuth: float
-    radius_m: Optional[float] = 600
+    azimuth: int
+    radius_m: Optional[int] = 600
     remark: Optional[str] = 'New Site'
     tlp_id: Optional[str] = 'N/A'
     tlp_name: Optional[str] = 'N/A'
